Move Surgery from shared to internal
