+++ ibl_pipeline/action_internal.py
@@ -1,5 +1,6 @@
 import datajoint as dj
 from . import reference, subject
+from .action_shared import ProcedureType
 import os
 
 mode = os.environ.get('MODE')
@@ -173,4 +174,39 @@ class Cull(dj.Manual):
     cull_method=null:               varchar(64)
     cull_description='':            varchar(1024)
     cull_ts=CURRENT_TIMESTAMP:      timestamp
+    """
+
+
+@schema
+class Surgery(dj.Manual):
+    # <class 'actions.models.Surgery'>
+    definition = """
+    -> subject.Subject
+    surgery_start_time:		datetime        # surgery start time
+    ---
+    surgery_uuid:           uuid
+    surgery_end_time=null:  datetime        # surgery end time
+    -> [nullable] reference.LabLocation.proj(surgery_lab='lab_name', surgery_location='location_name')
+    surgery_outcome_type:   enum('None', 'a', 'n', 'r')	    # outcome type
+    surgery_narrative=null: varchar(2048)	# narrative
+    surgery_ts=CURRENT_TIMESTAMP:   timestamp
+    """
+
+
+@schema
+class SurgeryUser(dj.Manual):
+    definition = """
+    -> Surgery
+    -> reference.LabMember
+    ---
+    surgeryuser_ts=CURRENT_TIMESTAMP:   timestamp
+    """
+
+
+@schema
+class SurgeryProcedure(dj.Manual):
+    definition = """
+    -> Surgery
+    -> ProcedureType
+    surgeryprocedure_ts=CURRENT_TIMESTAMP:   timestamp
     """
